Extract Merkle-Proof to IERC6956ValidAnchors
For collections, where the number of anchors is fixed (and a-priori
known), Serialized anchors may be used w/o security risks. In this case,
the check whether an anchor can be implemented simpler than with
merkle-trees, hence saving gas.

For the majority of use-cases, having an unknown and update-able number
of valid anchors is required, hence IERC6956ValidAnchors is
well-defined.